+++ x/data/server/query_attestations_by_attestor.go
@@ -1,0 +1,73 @@
+package server
+
+import (
+	"context"
+
+	"github.com/cosmos/cosmos-sdk/orm/model/ormlist"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
+	api "github.com/regen-network/regen-ledger/api/regen/data/v1"
+	"github.com/regen-network/regen-ledger/types"
+	"github.com/regen-network/regen-ledger/types/ormutil"
+	"github.com/regen-network/regen-ledger/x/data"
+)
+
+// AttestationsByAttestor queries attestations by attestor.
+func (s serverImpl) AttestationsByAttestor(ctx context.Context, request *data.QueryAttestationsByAttestorRequest) (*data.QueryAttestationsByAttestorResponse, error) {
+	if len(request.Attestor) == 0 {
+		return nil, sdkerrors.ErrInvalidRequest.Wrap("attestor cannot be empty")
+	}
+
+	addr, err := sdk.AccAddressFromBech32(request.Attestor)
+	if err != nil {
+		return nil, sdkerrors.ErrInvalidAddress.Wrap("attestor")
+	}
+
+	pg, err := ormutil.GogoPageReqToPulsarPageReq(request.Pagination)
+	if err != nil {
+		return nil, err
+	}
+
+	it, err := s.stateStore.DataAttestorTable().List(
+		ctx,
+		api.DataAttestorAttestorIndexKey{}.WithAttestor(addr),
+		ormlist.Paginate(pg),
+	)
+	if err != nil {
+		return nil, err
+	}
+
+	var attestations []*data.AttestationInfo
+	for it.Next() {
+		dataAttestor, err := it.Value()
+		if err != nil {
+			return nil, err
+		}
+
+		dataId, err := s.stateStore.DataIDTable().Get(ctx, dataAttestor.Id)
+		if err != nil {
+			return nil, err
+		}
+
+		dataAnchor, err := s.stateStore.DataAnchorTable().Get(ctx, dataAttestor.Id)
+		if err != nil {
+			return nil, err
+		}
+
+		attestations = append(attestations, &data.AttestationInfo{
+			Iri:       dataId.Iri,
+			Attestor:  request.Attestor,
+			Timestamp: types.ProtobufToGogoTimestamp(dataAnchor.Timestamp),
+		})
+	}
+
+	pageRes, err := ormutil.PulsarPageResToGogoPageRes(it.PageResponse())
+	if err != nil {
+		return nil, err
+	}
+
+	return &data.QueryAttestationsByAttestorResponse{
+		Attestations: attestations,
+		Pagination:   pageRes,
+	}, nil
+}
