+++ x/data/server/query_attestations_by_iri.go
@@ -1,0 +1,64 @@
+package server
+
+import (
+	"context"
+
+	"github.com/cosmos/cosmos-sdk/orm/model/ormlist"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
+
+	api "github.com/regen-network/regen-ledger/api/regen/data/v1"
+	"github.com/regen-network/regen-ledger/types"
+	"github.com/regen-network/regen-ledger/types/ormutil"
+	"github.com/regen-network/regen-ledger/x/data"
+)
+
+// AttestationsByIRI queries attestations to data by IRI.
+func (s serverImpl) AttestationsByIRI(ctx context.Context, request *data.QueryAttestationsByIRIRequest) (*data.QueryAttestationsByIRIResponse, error) {
+	if len(request.Iri) == 0 {
+		return nil, sdkerrors.ErrInvalidRequest.Wrap("iri cannot be empty")
+	}
+
+	dataId, err := s.stateStore.DataIDTable().GetByIri(ctx, request.Iri)
+	if err != nil {
+		return nil, sdkerrors.ErrNotFound.Wrapf("data entry with iri: %s", request.Iri)
+	}
+
+	pg, err := ormutil.GogoPageReqToPulsarPageReq(request.Pagination)
+	if err != nil {
+		return nil, err
+	}
+
+	it, err := s.stateStore.DataAttestorTable().List(
+		ctx,
+		api.DataAttestorIdAttestorIndexKey{}.WithId(dataId.Id),
+		ormlist.Paginate(pg),
+	)
+	if err != nil {
+		return nil, err
+	}
+
+	var attestations []*data.AttestationInfo
+	for it.Next() {
+		dataAttestor, err := it.Value()
+		if err != nil {
+			return nil, err
+		}
+
+		attestations = append(attestations, &data.AttestationInfo{
+			Iri:       request.Iri,
+			Attestor:  sdk.AccAddress(dataAttestor.Attestor).String(),
+			Timestamp: types.ProtobufToGogoTimestamp(dataAttestor.Timestamp),
+		})
+	}
+
+	pageRes, err := ormutil.PulsarPageResToGogoPageRes(it.PageResponse())
+	if err != nil {
+		return nil, err
+	}
+
+	return &data.QueryAttestationsByIRIResponse{
+		Attestations: attestations,
+		Pagination:   pageRes,
+	}, nil
+}
