--- x/data/server/query_resolvers.go
@@ -1,179 +1,0 @@
-package server
-
-import (
-	"context"
-
-	"github.com/cosmos/cosmos-sdk/orm/model/ormlist"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-	"github.com/cosmos/cosmos-sdk/types/query"
-
-	api "github.com/regen-network/regen-ledger/api/regen/data/v1"
-	"github.com/regen-network/regen-ledger/types/ormutil"
-	"github.com/regen-network/regen-ledger/x/data"
-)
-
-// ResolversByIRI queries all resolvers that have registered a piece of data by the IRI of the anchored data.
-func (s serverImpl) ResolversByIRI(ctx context.Context, request *data.QueryResolversByIRIRequest) (*data.QueryResolversByIRIResponse, error) {
-	dataId, err := s.stateStore.DataIDTable().GetByIri(ctx, request.Iri)
-	if err != nil {
-		return nil, sdkerrors.ErrNotFound.Wrap("content not anchored")
-	}
-
-	if request.Pagination == nil {
-		request.Pagination = &query.PageRequest{}
-	}
-
-	pg, err := ormutil.GogoPageReqToPulsarPageReq(request.Pagination)
-	if err != nil {
-		return nil, err
-	}
-
-	it, err := s.stateStore.DataResolverTable().List(
-		ctx,
-		api.DataResolverPrimaryKey{}.WithId(dataId.Id),
-		ormlist.Paginate(pg),
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	res := &data.QueryResolversByIRIResponse{}
-	for it.Next() {
-		item, err := it.Value()
-		if err != nil {
-			return nil, err
-		}
-
-		resolver, err := s.stateStore.ResolverTable().Get(ctx, item.ResolverId)
-		if err != nil {
-			return nil, err
-		}
-
-		manager := sdk.AccAddress(resolver.Manager).String()
-
-		res.Resolvers = append(res.Resolvers, &data.ResolverInfo{
-			Id:      resolver.Id,
-			Url:     resolver.Url,
-			Manager: manager,
-		})
-	}
-
-	res.Pagination, err = ormutil.PulsarPageResToGogoPageRes(it.PageResponse())
-	if err != nil {
-		return nil, err
-	}
-
-	return res, nil
-}
-
-// ResolversByHash queries all resolvers that have registered a piece of data by the ContentHash of the anchored data.
-func (s serverImpl) ResolversByHash(ctx context.Context, request *data.QueryResolversByHashRequest) (*data.QueryResolversByHashResponse, error) {
-	if request.ContentHash == nil {
-		return nil, sdkerrors.ErrInvalidRequest.Wrap("content hash cannot be empty")
-	}
-
-	iri, err := request.ContentHash.ToIRI()
-	if err != nil {
-		return nil, err
-	}
-
-	dataId, err := s.stateStore.DataIDTable().GetByIri(ctx, iri)
-	if err != nil {
-		return nil, sdkerrors.ErrNotFound.Wrap("content not anchored")
-	}
-
-	if request.Pagination == nil {
-		request.Pagination = &query.PageRequest{}
-	}
-
-	pg, err := ormutil.GogoPageReqToPulsarPageReq(request.Pagination)
-	if err != nil {
-		return nil, err
-	}
-
-	it, err := s.stateStore.DataResolverTable().List(
-		ctx,
-		api.DataResolverPrimaryKey{}.WithId(dataId.Id),
-		ormlist.Paginate(pg),
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	res := &data.QueryResolversByHashResponse{}
-	for it.Next() {
-		item, err := it.Value()
-		if err != nil {
-			return nil, err
-		}
-
-		resolver, err := s.stateStore.ResolverTable().Get(ctx, item.ResolverId)
-		if err != nil {
-			return nil, err
-		}
-
-		manager := sdk.AccAddress(resolver.Manager).String()
-
-		res.Resolvers = append(res.Resolvers, &data.ResolverInfo{
-			Id:      resolver.Id,
-			Url:     resolver.Url,
-			Manager: manager,
-		})
-	}
-
-	res.Pagination, err = ormutil.PulsarPageResToGogoPageRes(it.PageResponse())
-	if err != nil {
-		return nil, err
-	}
-
-	return res, nil
-}
-
-// ResolversByURL queries resolvers by resolver URL.
-func (s serverImpl) ResolversByURL(ctx context.Context, request *data.QueryResolversByURLRequest) (*data.QueryResolversByURLResponse, error) {
-	if len(request.Url) == 0 {
-		return nil, sdkerrors.ErrInvalidRequest.Wrap("url cannot be empty")
-	}
-
-	if request.Pagination == nil {
-		request.Pagination = &query.PageRequest{}
-	}
-
-	pg, err := ormutil.GogoPageReqToPulsarPageReq(request.Pagination)
-	if err != nil {
-		return nil, err
-	}
-
-	it, err := s.stateStore.ResolverTable().List(
-		ctx,
-		api.ResolverUrlIndexKey{}.WithUrl(request.Url),
-		ormlist.Paginate(pg),
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	res := &data.QueryResolversByURLResponse{}
-	for it.Next() {
-		item, err := it.Value()
-		if err != nil {
-			return nil, err
-		}
-
-		manager := sdk.AccAddress(item.Manager).String()
-
-		res.Resolvers = append(res.Resolvers, &data.ResolverInfo{
-			Id:      item.Id,
-			Url:     item.Url,
-			Manager: manager,
-		})
-	}
-
-	res.Pagination, err = ormutil.PulsarPageResToGogoPageRes(it.PageResponse())
-	if err != nil {
-		return nil, err
-	}
-
-	return res, nil
-}
