+++ x/data/server/query_resolvers_by_iri.go
@@ -1,0 +1,66 @@
+package server
+
+import (
+	"context"
+
+	"github.com/cosmos/cosmos-sdk/orm/model/ormlist"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
+	api "github.com/regen-network/regen-ledger/api/regen/data/v1"
+	"github.com/regen-network/regen-ledger/types/ormutil"
+	"github.com/regen-network/regen-ledger/x/data"
+)
+
+// ResolversByIRI queries resolvers with registered data by IRI.
+func (s serverImpl) ResolversByIRI(ctx context.Context, request *data.QueryResolversByIRIRequest) (*data.QueryResolversByIRIResponse, error) {
+	if len(request.Iri) == 0 {
+		return nil, sdkerrors.ErrInvalidRequest.Wrap("iri cannot be empty")
+	}
+
+	dataId, err := s.stateStore.DataIDTable().GetByIri(ctx, request.Iri)
+	if err != nil {
+		return nil, sdkerrors.ErrNotFound.Wrapf("data entry with iri: %s", request.Iri)
+	}
+
+	pg, err := ormutil.GogoPageReqToPulsarPageReq(request.Pagination)
+	if err != nil {
+		return nil, err
+	}
+
+	it, err := s.stateStore.DataResolverTable().List(
+		ctx,
+		api.DataResolverPrimaryKey{}.WithId(dataId.Id),
+		ormlist.Paginate(pg),
+	)
+	if err != nil {
+		return nil, err
+	}
+
+	res := &data.QueryResolversByIRIResponse{}
+	for it.Next() {
+		item, err := it.Value()
+		if err != nil {
+			return nil, err
+		}
+
+		resolver, err := s.stateStore.ResolverTable().Get(ctx, item.ResolverId)
+		if err != nil {
+			return nil, err
+		}
+
+		manager := sdk.AccAddress(resolver.Manager).String()
+
+		res.Resolvers = append(res.Resolvers, &data.ResolverInfo{
+			Id:      resolver.Id,
+			Url:     resolver.Url,
+			Manager: manager,
+		})
+	}
+
+	res.Pagination, err = ormutil.PulsarPageResToGogoPageRes(it.PageResponse())
+	if err != nil {
+		return nil, err
+	}
+
+	return res, nil
+}
